requirements.txt
click==6.6
Flask==0.11.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
nltk==3.2.1
oauthlib==2.0.0
pyStatParser==0.0.1
python-rtmidi==1.0.0
requests==2.11.1
requests-oauthlib==0.7.0
scipy==0.18.1
sh==1.11
six==1.10.0
twython==3.4.0
urllib3==1.19
vboxapi==1.0
virtualenv==1.11.6
Werkzeug==0.11.11
wget==3.2
